requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
filelock==3.13.1
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.4
keyboard==0.13.5
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.3
numba==0.61.0
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openai-whisper==20240930
pillow==11.0.0
pycparser==2.22
pyperclip==1.9.0
regex==2024.11.6
requests==2.32.3
setuptools==70.2.0
sounddevice==0.5.1
soundfile==0.13.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
